requirements.txt
pandas
python-dotenv
numpy
pymongo
flask
mlflow
dvc
Dockerfile
certifi
scikit-learn
scipy
dill
pyyaml
-e .